--- services/ui-src/src/components/CoreQuestions/index.tsx
@@ -1,4 +1,0 @@
-export * from "./AreYouReporting";
-export * from "./StatusOfDataReported";
-export * from "./CoreQuestionWrapper";
-export * from "./DataSource";
