+++ services/ui-src/src/views/DemoMeasure/DemoFormType.ts
@@ -1,0 +1,19 @@
+export namespace DemoForm {
+  export interface DemoFormType {
+    areYouReporting: string;
+    statusOfData: StatusOfData;
+    dataSource: DataSource;
+  }
+
+  export interface StatusOfData {
+    status: string;
+    statusOfDataAdditional?: string;
+  }
+
+  export interface DataSource {
+    "reporting.selections": string[];
+    "adminData.selections"?: string[];
+    "adminData.other.dataSource"?: string;
+    "adminData.other.explain"?: string;
+  }
+}
