--- services/ui-src/src/views/DemoQuestions/DemoFormType.ts
@@ -1,13 +1,0 @@
-export namespace DemoForm {
-  export interface DemoFormType {
-    areYouReporting: string;
-    statusOfDataReporting: StatusOfDataReportingNested;
-    dataSource: any;
-  }
-
-  export interface StatusOfDataReportingNested {
-    statusOfDataReporting: string;
-    statusOfDataReportingAdditional?: string;
-  }
-}
-
