+++ services/ui-src/src/views/index.tsx
@@ -9,4 +9,4 @@ export * from "./NotFound";
 export * from "./StateHome";
 export * from "./AddChildCoreSet";
 export * from "./AddHHCoreSet";
-export * from "./DemoQuestions";
+export * from "./DemoMeasure";
